Move `Validator` up to `rasa.validator`
